feat: 상세 페이지API  작업 및 동적 파라미터 추가
